# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Tilawat Al-Haramain Team Tasks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reciters:
    get:
      operationId: listReciters
      tags: [reciters]
      summary: List all reciters
      parameters:
        - name: mosque
          in: query
          required: false
          schema:
            type: string
            enum: [nabawi, haram]
      responses:
        "200":
          description: List of reciters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reciter"
    post:
      operationId: createReciter
      tags: [reciters]
      summary: Create a new reciter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReciterBody"
      responses:
        "201":
          description: Created reciter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reciter"
  /reciters/{id}:
    get:
      operationId: getReciter
      tags: [reciters]
      summary: Get a reciter by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reciter details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reciter"
    put:
      operationId: updateReciter
      tags: [reciters]
      summary: Update a reciter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReciterBody"
      responses:
        "200":
          description: Updated reciter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reciter"
    delete:
      operationId: deleteReciter
      tags: [reciters]
      summary: Delete a reciter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Reciter:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mosque:
          type: string
          enum: [nabawi, haram]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - mosque
        - createdAt
    CreateReciterBody:
      type: object
      properties:
        name:
          type: string
        mosque:
          type: string
          enum: [nabawi, haram]
      required:
        - name
        - mosque
...